+++ xxl-job-admin/src/main/java/com/xxl/job/admin/dao/XxlJobInfoDao.java
@@ -31,13 +31,13 @@ public interface XxlJobInfoDao {
 	
 	public int save(XxlJobInfo info);
 
-	public XxlJobInfo loadById(@Param("id") int id);
+	public XxlJobInfo loadById(@Param("id") long id);
 	
 	public int update(XxlJobInfo xxlJobInfo);
 	
 	public int delete(@Param("id") long id);
 
-	public List<XxlJobInfo> getJobsByGroup(@Param("jobGroup") int jobGroup);
+	public List<XxlJobInfo> getJobsByGroup(@Param("jobGroup") long jobGroup);
 
 	public int findAllCount();
